Added query dynamiclly
